Refactor: commonalize types of base title and base text components
