+++ components/BaseTitle.tsx
@@ -1,20 +1,11 @@
+import type { Props as BaseProps } from '@/types/BaseText'
 import BaseText from '@/components/BaseText'
 
-type Props = {
-  face?: 'normal' | 'ja' | 'en'
-  weight?: 'r' | 'm' | 'b'
-  className?: string
+type Props = BaseProps & {
   level: '1' | '2' | '3' | '4' | '5' | '6'
-  sizeSp: number
-  sizePc: number
-  children: React.ReactNode
 }
 
-const BaseTitle: React.VFC<Props> = ({
-  level,
-  children,
-  ...props
-}): JSX.Element => {
+const BaseTitle: React.VFC<Props> = ({ level, children, ...props }) => {
   return (
     <BaseText tagName={`h${level}`} {...props}>
       {children}
