+++ types/BaseText.ts
@@ -1,0 +1,10 @@
+type Props = {
+  face?: 'normal' | 'ja' | 'en'
+  weight?: 'r' | 'm' | 'b'
+  className?: string
+  sizeSp: number
+  sizePc: number
+  children: React.ReactNode
+}
+
+export type { Props }
